backend/requirements.txt
cloudinary
django==4.2
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-yasg
pillow
python-decouple
psycopg2